backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.13.0
python-dotenv==1.0.1
PyJWT==2.10.1
